use some enums

+++ TempusApi/Models/Activity/MapInfo.cs
@@ -4,7 +4,7 @@ namespace TempusApi.Models.Activity
     {
         [JsonPropertyName("date_added")] public double? DateAdded { get; set; }
 
-        [JsonPropertyName("id")] public int Id { get; set; }
+        [JsonPropertyName("id")] public long Id { get; set; }
 
         [JsonPropertyName("name")] public string Name { get; set; }
     }
