+++ TempusApi/Models/DetailedMapList/Author.cs
@@ -2,10 +2,10 @@
 {
     public class Author
     {
-        [JsonPropertyName("map_id")] public int MapId { get; set; }
+        [JsonPropertyName("map_id")] public long MapId { get; set; }
 
         [JsonPropertyName("name")] public string Name { get; set; }
 
-        [JsonPropertyName("id")] public int Id { get; set; }
+        [JsonPropertyName("id")] public long Id { get; set; }
     }
 }
