+++ TempusApi/Models/PlayerStats/BaseZoneStats.cs
@@ -2,7 +2,7 @@
 {
     public class BaseZoneStats
     {
-        [JsonPropertyName("count")] public int Count { get; set; }
+        [JsonPropertyName("count")] public long Count { get; set; }
 
         [JsonPropertyName("points")] public double Points { get; set; }
     }
