+++ TempusApi/Models/Responses/RanksOverviewModel.cs
@@ -4,7 +4,7 @@ namespace TempusApi.Models.Responses
 {
     public class RanksOverviewModel
     {
-        [JsonPropertyName("count")] public int PlayerCount { get; set; }
+        [JsonPropertyName("count")] public long PlayerCount { get; set; }
 
         [JsonPropertyName("players")] public List<PlayerRankOverviewModel> TopPlayers { get; set; }
     }
