+++ TempusApi/Models/Responses/RecordZoneInfo.cs
@@ -2,9 +2,9 @@
 {
     public class RecordZoneInfo
     {
-        [JsonPropertyName("id")] public int Id { get; set; }
+        [JsonPropertyName("id")] public long Id { get; set; }
 
-        [JsonPropertyName("map_id")] public int MapId { get; set; }
+        [JsonPropertyName("map_id")] public long MapId { get; set; }
 
         [JsonPropertyName("type")] public string Type { get; set; }
