+++ TempusApi/Models/ServerPlayerModel.cs
@@ -6,6 +6,6 @@
 
         [JsonPropertyName("name")] public string Name { get; set; }
 
-        [JsonPropertyName("id")] public int? Id { get; set; }
+        [JsonPropertyName("id")] public long? Id { get; set; }
     }
 }
